+++ app/services/parsers/ztParser.js
@@ -1,15 +1,12 @@
-const helpers = require('../../helpers/helpers')
-const ztService = require('../zieloneTarasyService')
+const wixCrawler = require('../wixCrawler')
 
 const ztParser = $ => {
-  const page = $.html()
-  return ztService.getZieloneTarasyPageObj(page)
-    .then(pageObj => {
-      const menuObj = JSON.parse(pageObj)
-      const menuHTML = menuObj['data']['document_data']['c2pd']['text']
+  return wixCrawler.crawlWix('https://www.zielone-tarasy.eu/', 'aktualności').then(pageObj => {
+    const menuObj = JSON.parse(pageObj)
+    const menuHTML = menuObj['data']['document_data']['c2pd']['text']
 
-      return helpers.buildTarasyMenu(stripHTML(menuHTML))
-    })
+    return buildTarasyMenu(stripHTML(menuHTML))
+  })
 }
 
 const stripHTML = html => {
@@ -21,6 +18,14 @@ const stripHTML = html => {
   ]
   const re = new RegExp(tagMap.join('|'), 'g')
   return html.replace(re, '').replace(/\r?\n|\r/g, ' ')
+}
+
+const buildTarasyMenu = (html) => {
+  return [
+    html.substring(html.indexOf('VEGE'), html.indexOf('M&B')),
+    html.substring(html.indexOf('M&B'), html.indexOf('TRZECI')),
+    html.substring(html.indexOf('TRZECI'))
+  ]
 }
 
 module.exports = ztParser
