+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "body-parser": "~1.17.1",
     "cheerio": "^1.0.0-rc.1",
+    "co": "^4.6.0",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.3",
     "express": "~4.15.2",
